# Conflicts:
#	engine/NetStart.go
#	state/authority.go
#	state/identityUtils.go
#	state/saveAndRestore.go

+++ controlPanel/dataDumpFormatting/servers.go
@@ -68,7 +68,7 @@ func MyNodeInfo(copyDS state.DisplayState) string {
 	return prt
 }
 
-func returnStatString(i int) string {
+func returnStatString(i uint8) string {
 	var stat string
 	switch i {
 	case 0:
